+++ src/handlers/victorops/handler_victorops_out_util.go
@@ -1,0 +1,14 @@
+package victorops
+
+import (
+	cc "github.com/commonchat/commonchat-go"
+	"github.com/grokify/webhook-proxy-go/src/util"
+)
+
+func ExampleMessage(data util.ExampleData) (cc.Message, error) {
+	bytes, err := data.ExampleMessageBytes(HandlerKey, "formatted1")
+	if err != nil {
+		return cc.Message{}, err
+	}
+	return Normalize(bytes)
+}